# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ATai API Documentation",
        "description": "https://atai.org/api/v1.0",
        "termsOfService": "https://atai.org",
        "contact": {
            "email": "hotro@atai.org"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://atai.dev24h.net/api/v1.0",
            "description": "Live host"
        },
        {
            "url": "http://atai.org/api/v1.0",
            "description": "Local host"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "get all posts",
                "operationId": "getAllPosts",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "API Key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "total record return",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/posts/category/getAll": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "get all posts category",
                "operationId": "getAllPostsCategory",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "API Key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/posts/detail": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "get detail posts",
                "operationId": "getDetailPosts",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "API Key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID Tin tức",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
...
}